Re-organising Xml Suite related tests
